split up portallink tests and portal tests
