+++ Tests/UnitTests/Core/Geometry/CMakeLists.txt
@@ -32,4 +32,5 @@ add_unittest(TrapezoidVolumeBounds TrapezoidVolumeBoundsTests.cpp)
 add_unittest(VolumeBounds VolumeBoundsTests.cpp)
 add_unittest(Volume VolumeTests.cpp)
 add_unittest(CylinderVolumeStack CylinderVolumeStackTests.cpp)
+add_unittest(PortalLink PortalLinkTests.cpp)
 add_unittest(Portal PortalTests.cpp)
